refactor: updated regenerate and assemble tasks to accept and AuthoredObject type
In some instances AuthoredObject types are used outside of the scope of tasks.
Allowing the the tasks to directly take the AuthoredObject type allows these
configured objects to be reused. It also decouples the task from the model
specific AuthoredObject configuration.

BREAKING CHANGE: Breaking change to the public API for AssembleTask and RegenerateTask

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>