PSCE-302 feat: adds a task to sync third party content to a local trestle workspace (#137)
* chore: adds a check during AuthoredObject instantionation that the root is valid

AuthoredObjects need to exists in valid trestle workspace. Handling the
case that the working directory is not initialized with trestle.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: adds trestle init to tmp_repo fixture

Reduces code duplication in the end to end tests and any test that
has git and trestle checks

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat: initial implementation for SyncUpstreamsTask class

The adds one source type, git-managed, and copies models in
the supported trestle model list (e.g. catalogs, profiles) from a
git repository to a local trestle workspace.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: fixes comments in sync_upstreams_task.py

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: updates `fetch_oscal_content` based on PR feedback

It makes sense to move the exception handling closer to
where the exceptions are being thrown and allows this method
called in other places.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

---------

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>