PSCE-303 feat: adds trestlebot-sync-upstreams command (#142)
* feat(entrypoint): adds sync-upstreams entrypoint

E2E tests are updated to include a step with sync_upstreams
to ssp-authoring workflow

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: removes additional test data and modifies source test profile

We don't need to store an additional test file with small edits. This
updates the prep_upstream_repo function to just alter the profile and
write it out.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: incorporates changes based on PR feedback

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: updates include_models/exclude_models to model_names

Using the term include_models/exclude_models could leave confusion about what
is being filtered in this instance (model vs model instance). Updating
to model_names to be more precise.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

---------

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>