refactor: migrates rule validation to pydantic  (#207)
* feat: migrates all rule validation logic to pydantic

Pydantic custom rule validation and error handling with some
minor tweaks supports can accomplish what the goal was for the
ValidationHandler class. Removing in favor of a more cohesive
validation solution.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: moves all pydantic logic to trestle_rule

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: define alternative parameters as YAML in test data

To make the YAML test data more readable, define the alternative
paramerer mappings in proper YAML.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat: removes default string requirement in default value check

The default key value can be set automatically during transformation
to keep it from being defined twice in YAML. If defined twice, it will
just be validated that it matches the default value string.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat: removes exc info true by default on generic error handling

When the CLI errors, the error logs are not very clear and the error
should multiple times. This will help make the error more readable.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat: refines rule reporting format for more clarity

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* test: adds test case for a missing default value key

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: removes extra newline when processing validation errors

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat: centralizes rule validation error collection and formatting

For validation with pydantic per YAML file, we want to collect
all of the validation errors up front and give a consistent, detailed message
for an efficient workflow.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: adds the path instead of the name in rule transform error

The path contains the component information making the
file easier to indentify for fixing errors.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* refactor: refines logic in convert_errors

Renamed variable to be more descriptive and removes extra call to
errors() method

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

---------

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>