PSCE-408 feat: adds flags to set git provider information when interacting with the API (#217)
* feat: adds GitProviderFactory class

The GitProviderFactory class manages the creation
of the GitProvider based on input and default logic.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* refactor: organizes arguments by groups

As the arguments list grows, grouping will make
understanding how to use and when to set much
clearer

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat: adds arguments for providing git provider type and url

Using flag option to add to logic Git provider group for
useful help messages

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* docs: updates README.md with support Git provider information

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: updates logic to throw error with no with-token flag is set

The logic was checking if with_token is None or empty which is
never true. This change updates the with-token flag to set to
true to evaluate if standard input should be read.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat: adds an environment variable option for token

To simplify running locally in headless mode, an environment
variable would be suitable over piping to standard input. This
also simplifies the action entrypoint scripts.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* docs: adds guidance in action files around the "github_token" input

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: updates with-token flag to use const parameter

This returns the with-token flag to its original implementation
to avoid breaking changes, but uses the const instead of default parameter.
This will make sure sys.stdin is only set if the flag is present.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: updates entrypoint base to ignore empty target branch

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* docs: updates token input descriptions in actions files

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* docs: updates "update-actions-readmes" to handle newlines

Newlines in descriptions caused issues in table formatting

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* refactor: moves auto-detect logic to entrypoint base

All reusable input/output logic should be defined in
entrypoint_base.py for easier management and readability

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* docs: minor docstring grammar fixes

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: updates load_provider_from_environment to use configured logger

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

---------

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>